% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorize.heat}
\alias{colorize.heat}
\title{Colorize numerical values for the heat palette}
\usage{
colorize.heat(values, resolution = 4)
}
\arguments{
\item{resolution}{Limit resolution for small values}

\item{x}{Values to normalize}
}
\value{
Colorized values
}
\description{
Colorize numerical values for the heat palette
}
