% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{create.ionet}
\alias{create.ionet}
\title{Create interactive omics network objects}
\usage{
create.ionet(x, type = c("ig", "adj", "wadj", "onet"), ...)
}
\arguments{
\item{x}{Network data}

\item{type}{Type of input data}
}
\value{
A interactive omics networks object
}
\description{
Create interactive omics network objects
}
