% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{create.ionets}
\alias{create.ionets}
\title{Create a list of interactive omics network objects}
\usage{
create.ionets(x, type = c("ig", "adj", "wadj", "onet"), ...)
}
\arguments{
\item{x}{A list of network data}

\item{type}{Type of input data}
}
\value{
A list of interactive omics networks objects
}
\description{
Create a list of interactive omics network objects
}
