% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{ggwrs}
\alias{ggwrs}
\title{Weighted rank scores plot implement in ggplot}
\usage{
ggwrs(n, p = 1)
}
\arguments{
\item{n}{The number of ranks}

\item{p}{Weighting exponent}
}
\value{
A ggplot object
}
\description{
Weighted rank scores plot implement in ggplot
}
