% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{ig.betweeness}
\alias{ig.betweeness}
\title{Node-level betweeness centrality}
\usage{
ig.betweeness(ig)
}
\arguments{
\item{ig}{An igraph object}
}
\value{
A vector
}
\description{
Node-level betweeness centrality
}
