% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{ig.create}
\alias{ig.create}
\title{Helper function for creating graph objects}
\usage{
ig.create(data, type = c("ig", "adj", "wadj"))
}
\arguments{
\item{data}{Data object}

\item{type}{Type of object}
}
\value{
A list of named functions
}
\description{
Helper function for creating graph objects
}
