% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{ig.dangalchev}
\alias{ig.dangalchev}
\title{Node-level dangalchev closeness centrality}
\usage{
ig.dangalchev(ig)
}
\arguments{
\item{ig}{An igraph object}
}
\value{
A vector
}
\description{
Node-level dangalchev closeness centrality
}
