% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{ig.diffusion}
\alias{ig.diffusion}
\title{Node-level diffusion centrality}
\usage{
ig.diffusion(ig)
}
\arguments{
\item{ig}{An igraph object}
}
\value{
A vector
}
\description{
Node-level diffusion centrality
}
