% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{ig.eccentricity}
\alias{ig.eccentricity}
\title{Node-level eccentricity centrality}
\usage{
ig.eccentricity(ig)
}
\arguments{
\item{ig}{An igraph object}
}
\value{
A vector
}
\description{
Node-level eccentricity centrality
}
