% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{ig.laplacian}
\alias{ig.laplacian}
\title{Node-level laplacian centrality}
\usage{
ig.laplacian(ig)
}
\arguments{
\item{ig}{An igraph object}
}
\value{
A vector
}
\description{
Node-level laplacian centrality
}
