% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{ig.properties}
\alias{ig.properties}
\title{Functions for calculating graph-level properties}
\usage{
ig.properties()
}
\value{
A list of named functions
}
\description{
Functions for calculating graph-level properties
}
