% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ig}
\alias{ig}
\title{centrality}
\usage{
ig

ig
}
\description{
Derived from Zachary Karate Club Network

Simulated scale-free network through preferential attachment
}
\references{
W. W. Zachary, "An information flow model for conflict and fission in small groups." Journal of Anthropological Research 33, 452-473 (1977).
}
\keyword{datasets}
