% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propogation.R
\name{ig.rwr}
\alias{ig.rwr}
\title{Random walk with restart}
\usage{
ig.rwr(ig, seeds, restart = 0.5)
}
\arguments{
\item{ig}{An igraph object}

\item{seeds}{One or more seed node ids}

\item{restart}{Restart probability between 0-1}
}
\value{
A vector
}
\description{
Random walk with restart
}
