% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{ig.topoco}
\alias{ig.topoco}
\title{Node-level topological coefficient centrality}
\usage{
ig.topoco(ig)
}
\arguments{
\item{ig}{An igraph object}
}
\value{
A vector
}
\description{
Node-level topological coefficient centrality
}
