% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{interactive.omics.network}
\alias{interactive.omics.network}
\title{Interaction Omics Network Object}
\description{
Interaction Omics Network Object

Interaction Omics Network Object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ig}}{An igraph object}

\item{\code{pca}}{Dimensionality reduction on node measures}

\item{\code{nodes}}{A data frame of labeled nodes}

\item{\code{edges}}{A data frame of undirected edges}

\item{\code{properties}}{A list of graph-level properties
Init}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{interactive.omics.network$new()}}
\item \href{#method-print}{\code{interactive.omics.network$print()}}
\item \href{#method-init.properties}{\code{interactive.omics.network$init.properties()}}
\item \href{#method-init.nodes}{\code{interactive.omics.network$init.nodes()}}
\item \href{#method-init.edges}{\code{interactive.omics.network$init.edges()}}
\item \href{#method-get.symbols}{\code{interactive.omics.network$get.symbols()}}
\item \href{#method-get.interactors}{\code{interactive.omics.network$get.interactors()}}
\item \href{#method-get.subnetwork}{\code{interactive.omics.network$get.subnetwork()}}
\item \href{#method-plt.subnetwork}{\code{interactive.omics.network$plt.subnetwork()}}
\item \href{#method-clone}{\code{interactive.omics.network$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{interactive.omics.network$new(
  ig,
  symbols = NULL,
  graph.measures = c("nodes", "edges", "density", "clustering"),
  node.measures = c("degree", "eigen", "betweenness", "stress"),
  quiet = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ig}}{An undirected igraph object}

\item{\code{symbols}}{A mapping of node identifiers to symbols}

\item{\code{graph.measures}}{A vector of graph measures to compute}

\item{\code{node.measures}}{A vector of node measures to compute}

\item{\code{quiet}}{Use true to silence messages}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new network object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print network information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{interactive.omics.network$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init.properties"></a>}}
\if{latex}{\out{\hypertarget{method-init.properties}{}}}
\subsection{Method \code{init.properties()}}{
Computes graph-level properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{interactive.omics.network$init.properties(measures = c(), quiet = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measures}}{A vector of measures to compute}

\item{\code{quiet}}{Use true to silence messages}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init.nodes"></a>}}
\if{latex}{\out{\hypertarget{method-init.nodes}{}}}
\subsection{Method \code{init.nodes()}}{
Computes node-level properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{interactive.omics.network$init.nodes(
  symbols = NULL,
  measures = c(),
  quiet = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{A mapping of node identifers to symbols}

\item{\code{measures}}{A vector of measures to compute}

\item{\code{quiet}}{Use true to silence messages}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init.edges"></a>}}
\if{latex}{\out{\hypertarget{method-init.edges}{}}}
\subsection{Method \code{init.edges()}}{
Computes edge-level properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{interactive.omics.network$init.edges()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.symbols"></a>}}
\if{latex}{\out{\hypertarget{method-get.symbols}{}}}
\subsection{Method \code{get.symbols()}}{
Maps ids to symbols
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{interactive.omics.network$get.symbols(ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{A vector of node identifiers}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of symbols
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.interactors"></a>}}
\if{latex}{\out{\hypertarget{method-get.interactors}{}}}
\subsection{Method \code{get.interactors()}}{
Get interactors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{interactive.omics.network$get.interactors(
  ids,
  degree = 1,
  remove.ids = TRUE,
  use.symbols = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{Node identifiers}

\item{\code{degree}}{Edge degree to include in subset}

\item{\code{remove.ids}}{Include original ids}

\item{\code{use.symbols}}{Return node symbols instead of ids}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of interactors
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.subnetwork"></a>}}
\if{latex}{\out{\hypertarget{method-get.subnetwork}{}}}
\subsection{Method \code{get.subnetwork()}}{
Get subnetwork
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{interactive.omics.network$get.subnetwork(
  ids,
  degree = 1,
  indirect.edges = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{Node identifiers}

\item{\code{degree}}{Edge degree to include in subset}

\item{\code{indirect.edges}}{Include edges between non-queried nodes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of nodes and edges
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plt.subnetwork"></a>}}
\if{latex}{\out{\hypertarget{method-plt.subnetwork}{}}}
\subsection{Method \code{plt.subnetwork()}}{
Plot subnetwork
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{interactive.omics.network$plt.subnetwork(
  ids,
  degree = 1,
  indirect.edges = FALSE,
  use.symbols = FALSE,
  node.color = NULL,
  node.border = NULL,
  node.size = NULL,
  node.shape = NULL,
  layout = "layout_nicely",
  node.attr = list(),
  edge.attr = list(),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{Node identifiers}

\item{\code{degree}}{Edge degree to include in subset}

\item{\code{indirect.edges}}{Include edges between non-queried nodes}

\item{\code{use.symbols}}{Show node symbols instead of ids}

\item{\code{node.color}}{Node column to determine node background color}

\item{\code{node.border}}{Node column to determine node border color}

\item{\code{node.size}}{Node column to determine node size}

\item{\code{node.shape}}{Node column to determine node shape}

\item{\code{layout}}{An igraph layout}

\item{\code{node.attr}}{Node attributes passed to `visNodes()`}

\item{\code{edge.attr}}{Edge attributes passed to `visEdges()`}

\item{\code{...}}{Additional keyword arguments passed to `visNetwork()`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A visNetwork object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{interactive.omics.network$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
