% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{is.network}
\alias{is.network}
\title{Checks if object is a interactive omics network object}
\usage{
is.network(network)
}
\arguments{
\item{network}{An object}
}
\value{
A bool
}
\description{
Checks if object is a interactive omics network object
}
