% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{is.networks}
\alias{is.networks}
\title{Checks if object is a list of interactive omics network objects}
\usage{
is.networks(networks)
}
\arguments{
\item{networks}{A list of objects}
}
\value{
A bool
}
\description{
Checks if object is a list of interactive omics network objects
}
