% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstest.R
\name{kstest}
\alias{kstest}
\title{Two-sided Kolmogorov–Smirnov test}
\usage{
kstest(
  n.x,
  y,
  weights = NULL,
  weights.pwr = 1,
  absolute = FALSE,
  plotting = FALSE
)
}
\arguments{
\item{n.x}{The length of a ranked list}

\item{y}{A vector of positions in the ranked list}

\item{weights}{Weights for weighted score (Subramanian et al.)}

\item{weights.pwr}{Exponent for weights (Subramanian et al.)}

\item{absolute}{Takes max-min score rather than the max deviation from null}

\item{plotting}{Use true to generate plot}
}
\value{
A list of data and plots
}
\description{
Two-sided Kolmogorov–Smirnov test
}
