% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{network.kstest}
\alias{network.kstest}
\title{Centrality-based enrichment}
\usage{
network.kstest(network, metric, geneset)
}
\arguments{
\item{network}{An interactive omics network object}

\item{metric}{Centrality measure}

\item{geneset}{A list of genes or other omics features}
}
\value{
A ggplot object
}
\description{
Centrality-based enrichment
}
