% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{network.pca.contri}
\alias{network.pca.contri}
\title{PCA - Contibutors}
\usage{
network.pca.contri(network)
}
\arguments{
\item{network}{An interactive omics network object}
}
\value{
A ggplot object
}
\description{
PCA - Contibutors
}
