% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{network.pca.hclust}
\alias{network.pca.hclust}
\title{PCA - Hierarchical Clustering}
\usage{
network.pca.hclust(network)
}
\arguments{
\item{network}{An interactive omics network object}
}
\value{
A ggplot object
}
\description{
PCA - Hierarchical Clustering
}
