% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{network.pca.pltvar}
\alias{network.pca.pltvar}
\title{PCA - Show Contributors}
\usage{
network.pca.pltvar(network)
}
\arguments{
\item{network}{An interactive omics network object}
}
\value{
A ggplot object
}
\description{
PCA - Show Contributors
}
