% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{network.pca.varexp}
\alias{network.pca.varexp}
\title{PCA - Variance Explained}
\usage{
network.pca.varexp(network)
}
\arguments{
\item{network}{An interactive omics network object}
}
\value{
A ggplot object
}
\description{
PCA - Variance Explained
}
