% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{networks.cdistr}
\alias{networks.cdistr}
\title{Centrality distributions}
\usage{
networks.cdistr(
  networks,
  metric,
  plot = c("boxplot", "density", "ecdf"),
  rescale = FALSE
)
}
\arguments{
\item{networks}{A named list of interactive omics network objects}

\item{metric}{Centrality measure}

\item{plot}{Plot type}

\item{rescale}{Rescale values between 0-1}
}
\value{
A ggplot object
}
\description{
Centrality distributions
}
