% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{networks.diffc}
\alias{networks.diffc}
\title{Differential Centrality Rankings Across Networks}
\usage{
networks.diffc(networks, metric, p = 1, top = 25, show = c("r", "v"))
}
\arguments{
\item{networks}{A named list of interactive omics network objects}

\item{metric}{Centrality measure}

\item{p}{Scaling exponent}

\item{top}{Number of nodes to include}

\item{show}{Show rank or metric value}
}
\value{
A ggplot object
}
\description{
Differential Centrality Rankings Across Networks
}
