% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{networks.hclust}
\alias{networks.hclust}
\title{Hierarchical Clustering of Networks by Centrality}
\usage{
networks.hclust(networks, metric, method = c("pearson", "kendall"))
}
\arguments{
\item{networks}{A named list of interactive omics network objects}

\item{metric}{Centrality measure}

\item{method}{Correlation method}
}
\value{
A ggplot object
}
\description{
Hierarchical Clustering of Networks by Centrality
}
