% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize.nlog.range}
\alias{normalize.nlog.range}
\title{Normalize significance values between a given range}
\usage{
normalize.nlog.range(x, base = 10, a = 0, b = 1)
}
\arguments{
\item{x}{Values to normalize}

\item{base}{Base of logarithm}

\item{a}{Min of range}

\item{b}{max of range}
}
\value{
Normalized values
}
\description{
Normalize significance values between a given range
}
