% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize.nlog}
\alias{normalize.nlog}
\title{Normalize significance values}
\usage{
normalize.nlog(x, base = 10)
}
\arguments{
\item{x}{Values to normalize}

\item{base}{Base of logarithm}
}
\value{
Normalized values

Normalized values
}
\description{
Normalize significance values
}
