% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize.zo}
\alias{normalize.zo}
\title{Normalize values between zero and one}
\usage{
normalize.zo(x)
}
\arguments{
\item{x}{Values to normalize}
}
\value{
Normalized values
}
\description{
Normalize values between zero and one
}
