% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pseudo.log.trans}
\alias{pseudo.log.trans}
\title{Pseudo-log transformation}
\usage{
pseudo.log.trans(sigma = 1, base = exp(1))
}
\arguments{
\item{sigma}{Scaling factor}

\item{base}{Base of logarithm}
}
\value{
A scale
}
\description{
Derived from the R package scales by Hadley Wickham
https://github.com/r-lib/scales
See scales::pseudo_log_trans()
}
