% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcolors}
\alias{rcolors}
\title{Repeatable vector of distinct colors}
\usage{
rcolors(reps = 1)
}
\arguments{
\item{reps}{Number of copies}
}
\value{
A vector of distinct values
}
\description{
Repeatable vector of distinct colors
}
