% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run.shiny}
\alias{run.shiny}
\title{Spin up an instance of the shiny application}
\usage{
run.shiny(launch.browser = TRUE, dev = TRUE)
}
\arguments{
\item{launch.browser}{Launch in browser mode}

\item{dev}{Refresh server upon code changes}
}
\description{
Spin up an instance of the shiny application
}
