% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{theme_simplex}
\alias{theme_simplex}
\title{An ggplot theme for simplex applications}
\usage{
theme_simplex(dark = FALSE)
}
\arguments{
\item{dark}{Background color}
}
\value{
A ggplot object
}
\description{
An ggplot theme for simplex applications
}
