% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{vn.to.ig}
\alias{vn.to.ig}
\title{Convert vn to ig}
\usage{
vn.to.ig(nodes, edges)
}
\arguments{
\item{nodes}{A dataframe of nodes}

\item{edges}{A dataframe of edges}
}
\value{
An ig object
}
\description{
Convert vn to ig
}
